CODY-89 move documentation files under documentation folder
